autoupdate/requirements.txt
pyyaml
graphqlclient
python-dotenv
